Enhanced Box-Percentile Plots.

Please use function BoxPlot with box.type="bp" for box-percentile plots, since this function here was designed as helper function for function BoxPlot. It uses the R-code of function bpplot which generates box-percentile plots as implemented in package Hmisc. This R-code is changed and augmented to implement an formula-interface and to be able to add box-percentile plots to an existing plot. It is assumed that form is a simple formula object, e.g. y~grp:time with just simple interactions or even simpler with just a single grouping factor, e.g. y~grp.

BPplot(obj, form = NULL, var = NULL, main = "Box-Percentile Plot",
  add = FALSE, col = NULL, labels = NULL, ylab = NULL,
  ylab.line = 2.5, ylab.font = 1L, swag = 0.9, line.col = "black",
  line.lwd = 1L, line.lty = 1L, add.xlab = TRUE, ...)

Arguments

obj
(data.frame, matrix) corresponding to the dataset to be used

form
(formula) such as y~grp, where y is a numeric vector of data values to be split into groups according to the grouping variable grp.

var
(character) vector specifying the columns in 'obj' to be used for plotting. The order of elements is retained in the boxplot.

main
(character) string giving the main title of the plot, left out if add=TRUE

add
(logical) TRUE=add box-percentile plot(s) to an existing plot

col
(character) vector specifying the colors of percentile-boxes, defaults to no color

labels
(character) string(s) for group labels, which are drawn under each boxplot. Note: The order must correspond to the order of group-levels if the formula interface was used, i.e. check the order of sort(unique(obj$grp)).

ylab
(character) character string specifying the optional label of the Y-axis (vertically centered)

ylab.line
(numeric) specifying the line 'ylab' is put (useful for custom figure margins).

ylab.font
(integer) specifying the font to be used for the Y-axis label

swag
(numeric) values within ]0,1] specifying the swaging of percentile boxes, for better separating boxes from each other

line.col
(character) color of boundry lines of the percentile boxes

line.lwd
(integer) line width of the boundry lines of the percentile boxes

line.lty
(integer) line type of the boundry lines of the percentile boxes

add.xlab
(logical) TRUE = automically determined group-labels are plotted below each box/percentile-box

...
additional graphical parameters passed on

See also

BoxPlot, function bpplot in package Hmisc