Log rank test p value and other summary statistics for subgroup analysis

This function takes time to event outcome, and one categorical variable (parameter var). It generates log rank p value comparing different groups, and summary statistics

LogRankTab(data, tte, cens, var, time.unit = "month",
  surv.conf.type = "plain", ties = "efron", fillname = "")

Arguments

data
Input data frame. Rows are patients and columns are variables (e.g. demographics variables, time to event variables, biomarker variables, treatment indicator, etc.). One patient per row.

tte
column name that indicates the time to event variable

cens
column name that indicates the censoring variable associated to tte. 1 indicates event and 0 indicates censoring

var
a vector that indicates variables of interest. Hazard ratio, CI and pvalue of these variables will be calculated. All elements in var should be also in input data's column names. It will be ignored if fit or form is not null.

time.unit
Default is month.

surv.conf.type
confidence interval type. Default is "plain". see conf.type in survfit

ties
Default is "efron". To match internal sas results, use "exact". See parameter "ties" in coxph.

fillname
(character) variable specifying the content for the cell that located in row one and column one, the default value for fillname is NULL.

Note

This function takes time to event outcome, and one categorical variable (parameter var). This function generates a table that contains number of patients, median time to event, and log rank test p value ( against first level in var) For more general time to event modeling, see function CoxTab() (e.g. continuous factor, multiple factors, stratification, hazard ratio estimation, etc)

Examples

data(input)
LogRankTab(data=input,tte="PFS", cens="PFS.event",var="Arm")

#>       [,1]                           [,2]            [,3]          
#>  [1,] ""                             "CTRL"          "TRT"         
#>  [2,] ""                             "N=182"         "N=368"       
#>  [3,] "Patients with event"          "120 (65.9%)"   "202 (54.9%)" 
#>  [4,] "Patients without event"       "62 (34.1%)"    "166 (45.1%)" 
#>  [5,] "Time to event (month)"        ""              ""            
#>  [6,] "     Median (KM)"             "3.25"          "5.82"        
#>  [7,] "     95% CI Median"           "(2.83;4.14)"   "(5.49;6.8)"  
#>  [8,] "     25% and 75%-ile"         "1.61;6.47"     "2.89;11.73"  
#>  [9,] "     Range (inc. cens.)"      "0.03 to 12.32" "0.03 to 13.8"
#> [10,] "     p-value (Log-Rank Test)" " "             "0"           
#> [11,] "Hazard Ratio"                 " "             "0.54"        
#> [12,] " 95% CI"                      " "             "(0.43;0.67)"