Choose an R standard color or get the RGB-code of an arbitrary color.

Function lets the user choose a color via a tcltk widget in case it is called with default value of paramater 'continuous'. Otherwise, it plots a grid of differently colored tiles, which can be used to select as many colors as desired by simply clicking at the corresponding tiles.

get.colors(continuous = TRUE)

Arguments

continuous
(logical) TRUE = one or multiple colors can be chosen from the RGB color-space via a tcltk-widget FALSE = a grid is drawn where R standard colors can be selected by clicking on colored-tiles

Examples

## get R standard color
 get.colors(FALSE)

#> Error in rect(i - 1, j - 1, i, j, col = Colors[(i - 1) * 26 + j]): plot.new has not been called yet

## get RGB-code (hexadecimal) via tcltk-widget
 
# NOT RUN {
     get.colors()

# }