Construct Matrix from Formula and Names.

Function is a helper function for adding tables as X-axis labels representing the structure of crossed (nested) factor-levels.

getMatrix(form, labels, split="\\.")

Arguments

form
(formula) object from which the matrix should be constructed

labels
(character) the 'names" element of the list returned by e.g. BoxPlot or boxplot

split
(character) symbol or string specifying the split symbol or string separating factor-levels in 'labels'

Examples

bp  <- boxplot(mpg~cyl:gear, mtcars, plot=FALSE)
mat <- getMatrix(mpg~cyl:gear, bp$names)
mat

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9]
#> cyl  "4"  "6"  "8"  "4"  "6"  "8"  "4"  "6"  "8" 
#> gear "3"  "3"  "3"  "4"  "4"  "4"  "5"  "5"  "5"