Get model formula from a model object

Search within a model object for model formula information and derive the formula If the model object contains a named element matching any name in the params parameter, that item will be returned, presumably containing the environment information in which it was instantiated.

If the model object doesn't store this information, the model call will be inspected to search for the formula argument used and derive formula from a new formula produced from this input. As a last resort, it will take the first passed argument and try to coerce it to a formula. A formula derived in this way will be created in the local function environment and will not be able to be used to evaluate the symbols stored within them.

get_model_formula(model.obj, params = c("terms", "model", "formula"))

Arguments

model.obj
the model object with which to derive the formula

params
model attributes or parameters to consder for deriving formula

Value

A formula object or item which may be coerced to a formula