Compute emmeans to mirror SAS emmeans functionality

Compute emmeans to mirror SAS emmeans functionality

sas.emmeans(model.obj, specs, data = NULL, mode = "kenward-roger",
  quietly = FALSE, verbose = FALSE, confidence.level = NULL, ...)

Arguments

model.obj
A model object on which to calculate emmeans

specs
specs to be passed to emmeans()

data
specified data to be used for model fitting. (defaults to data bound to model, though this can not always been inferred from the model object and scope of function call)

mode
mode to be passed to emmeans()

quietly
whether to hide message informing of changes to default functionality

verbose
whether all debug information should be printed to console

confidence.level
the confidence interval to use for upper and lower confidence level outputs

...
additional arguments to be passed to emmeans

Examples

library(dplyr) # for %>% operator
library(nlme)  # for gls

#> 
#> Attaching package: ‘nlme’

#> The following object is masked from ‘package:gClinBiomarker’:
#> 
#>     collapse

#> The following object is masked from ‘package:dplyr’:
#> 
#>     collapse

# create model on which to calculate emmeans
model <- gls(mpg ~ hp + carb + wt,
  data = mtcars, na.action = na.exclude)

# call emmeans, specifying factors over which means should be calculated
sas.emm <- sas.emmeans(model, ~ carb, data = mtcars, confidence.level = 0.9)

#> To mirror SAS functionality:
#>  1. mode of 'kenward-roger' will be used unless otherwise specified.
#>  2. data used for initial model fit...
#>    a. is filtered for only complete cases in the response variable
#>    b. will have any numeric variables in specs converted to categorical
#>  3. grouping variables for emmeans converted to factor.
#>       (creating a new factor column where necessary)
#>  4. continuous means calculated using rows where response is not NA.
#>  5. categorical level weights from all rows of dataset

#> 
#> Attaching package: ‘emmeans’

#> The following object is masked from ‘package:devtools’:
#> 
#>     test

# update the confidence level used
# NOTE: if you're examining confidence intervals, it's much faster to udpate
# rather than re-run the model and means
sas.emm <- update(sas.emm, level = 0.95)

# using the means as a dataframe
as.data.frame(sas.emm)

#>   carb carb.emmeans.factor   emmean        SE df lower.CL upper.CL .wgt.
#> 1    1                   1 20.63151 1.2682590 24 18.01395 23.24906     7
#> 2    2                   2 20.10695 0.9417013 24 18.16337 22.05053    10
#> 3    3                   3 19.74925 1.6608191 24 16.32148 23.17701     3
#> 4    4                   4 19.61794 1.0162966 24 17.52041 21.71548    10
#> 5    6                   6 19.11064 2.8433240 24 13.24231 24.97897     1
#> 6    8                   8 22.87216 3.7432900 24 15.14639 30.59793     1