Compute LSMEANS to mirror SAS functionality

Compute LSMEANS to mirror SAS functionality

sas.lsmeans(model.obj, specs, data = NULL, mode = "kenward-roger",
  quietly = FALSE, verbose = FALSE, confidence.level = NULL, ...)

Arguments

model.obj
A model object on which to calculate lsmeans

specs
specs to be passed to lsmeans()

mode
mode to be passed to lsmeans()

quietly
whether to hide message informing of changes to default functionality

...
additional arguments to be passed to lsmeans

as.data.frame
whether to return a lsm.list object or a data.frame

Examples

library(dplyr) # for %>% operator
library(nlme)  # for gls

#> 
#> Attaching package: ‘nlme’

#> The following object is masked from ‘package:dplyr’:
#> 
#>     collapse

#> The following object is masked from ‘package:gClinBiomarker’:
#> 
#>     collapse

# create model on which to calculate lsmeans
model <- gls(mpg ~ hp + carb + wt,
  data = mtcars, na.action = na.exclude)

# call lsmeans, specifying factors over which means should be calculated
sas.lsm <- sas.lsmeans(model, ~ carb, data = mtcars, confidence.level = 0.9, verb = T)

#> To mirror SAS functionality:
#>  1. mode of 'kenward-roger' will be used unless otherwise specified.
#>  2. data used for initial model fit...
#>    a. is filtered for only complete cases in the response variable
#>    b. will have any numeric variables in specs converted to categorical
#>  3. grouping variables for lsmeans converted to factor.
#>       (creating a new factor column where necessary)
#>  4. continuous means calculated using rows where response is not NA.
#>  5. categorical level weights from all rows of dataset

#> gls model is being refit with arguments: 
#> $model
#> mpg ~ hp + carb + wt
#> $data
#> mtcars
#> $na.action
#> na.exclude

#> Running lsmeans::lsmeans() with parameters: 

#> $object
#> Generalized least squares fit by REML
#>   Model: mpg ~ hp + carb + wt 
#>   Data: structure(list(mpg = c(21, 21, 22.8, 21.4, 18.7, 18.1, 14.3,  24.4, 22.8, 19.2, 17.8, 16.4, 17.3, 15.2, 10.4, 10.4, 14.7, 32.4,  30.4, 33.9, 21.5, 15.5, 15.2, 13.3, 19.2, 27.3, 26, 30.4, 15.8,  19.7, 15, 21.4), cyl = c(6, 6, 4, 6, 8, 6, 8, 4, 4, 6, 6, 8,  8, 8, 8, 8, 8, 4, 4, 4, 4, 8, 8, 8, 8, 4, 4, 4, 8, 6, 8, 4),      disp = c(160, 160, 108, 258, 360, 225, 360, 146.7, 140.8,      167.6, 167.6, 275.8, 275.8, 275.8, 472, 460, 440, 78.7, 75.7,      71.1, 120.1, 318, 304, 350, 400, 79, 120.3, 95.1, 351, 145,      301, 121), hp = c(110, 110, 93, 110, 175, 105, 245, 62, 95,      123, 123, 180, 180, 180, 205, 215, 230, 66, 52, 65, 97, 150,      150, 245, 175, 66, 91, 113, 264, 175, 335, 109), drat = c(3.9,      3.9, 3.85, 3.08, 3.15, 2.76, 3.21, 3.69, 3.92, 3.92, 3.92,      3.07, 3.07, 3.07, 2.93, 3, 3.23, 4.08, 4.93, 4.22, 3.7, 2.76,      3.15, 3.73, 3.08, 4.08, 4.43, 3.77, 4.22, 3.62, 3.54, 4.11     ), wt = c(2.62, 2.875, 2.32, 3.215, 3.44, 3.46, 3.57, 3.19,      3.15, 3.44, 3.44, 4.07, 3.73, 3.78, 5.25, 5.424, 5.345, 2.2,      1.615, 1.835, 2.465, 3.52, 3.435, 3.84, 3.845, 1.935, 2.14,      1.513, 3.17, 2.77, 3.57, 2.78), qsec = c(16.46, 17.02, 18.61,      19.44, 17.02, 20.22, 15.84, 20, 22.9, 18.3, 18.9, 17.4, 17.6,      18, 17.98, 17.82, 17.42, 19.47, 18.52, 19.9, 20.01, 16.87,      17.3, 15.41, 17.05, 18.9, 16.7, 16.9, 14.5, 15.5, 14.6, 18.6     ), vs = c(0, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 0,      0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1), am = c(1,      1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1,      0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1), gear = c(4, 4, 4, 3,      3, 3, 3, 4, 4, 4, 4, 3, 3, 3, 3, 3, 3, 4, 4, 4, 3, 3, 3,      3, 3, 4, 5, 5, 5, 5, 5, 4), carb = structure(c(4L, 4L, 1L,      1L, 2L, 1L, 4L, 2L, 2L, 4L, 4L, 3L, 3L, 3L, 4L, 4L, 4L, 1L,      2L, 1L, 1L, 2L, 2L, 4L, 2L, 1L, 2L, 2L, 4L, 5L, 6L, 2L), .Label = c("1",      "2", "3", "4", "6", "8"), class = "factor")), class = "data.frame", .Names = c("mpg",  "cyl", "disp", "hp", "drat", "wt", "qsec", "vs", "am", "gear",  "carb"), row.names = c(NA, -32L)) 
#>        AIC      BIC    logLik
#>   156.0622 166.6646 -69.03108
#> 
#> Coefficients:
#>                Value Std.Error   t-value p-value
#> (Intercept) 37.19123  1.916554 19.405259  0.0000
#> hp          -0.03517  0.014147 -2.485645  0.0203
#> carb2       -0.52456  1.412113 -0.371469  0.7135
#> carb3       -0.88226  2.228159 -0.395959  0.6956
#> carb4       -1.01356  1.830877 -0.553595  0.5850
#> carb6       -1.52087  3.173535 -0.479234  0.6361
#> carb8        2.24066  4.309517  0.519932  0.6079
#> wt          -3.54383  0.773606 -4.580932  0.0001
#> 
#>  Correlation: 
#>       (Intr) hp     carb2  carb3  carb4  carb6  carb8 
#> hp    -0.075                                          
#> carb2 -0.247 -0.199                                   
#> carb3  0.099 -0.332  0.480                            
#> carb4  0.135 -0.449  0.596  0.612                     
#> carb6 -0.105 -0.359  0.325  0.296  0.377              
#> carb8  0.055 -0.709  0.348  0.413  0.537  0.344       
#> wt    -0.652 -0.557 -0.030 -0.143 -0.160  0.153  0.261
#> 
#> Standardized residuals:
#>                      Min                       Q1                      Med 
#> -1.453000733540340760896 -0.518130806197143023439  0.000000000000001605611 
#>                       Q3                      Max 
#>  0.319725705209169319687  2.007393207317980987625 
#> 
#> Residual standard error: 2.765858 
#> Degrees of freedom: 32 total; 24 residual

#> $specs
#> pairwise ~ carb
#> <environment: 0x11a9328a0>
#> 
#> $mode
#> [1] "kenward-roger"
#> 
#> $data
#>     mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> 1  21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> 2  21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> 3  22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> 4  21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> 5  18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> 6  18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> 7  14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> 8  24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> 9  22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> 10 19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> 11 17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> 12 16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> 13 17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> 14 15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> 15 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> 16 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> 17 14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> 18 32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> 19 30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> 20 33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> 21 21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> 22 15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> 23 15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> 24 13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> 25 19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> 26 27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> 27 26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> 28 30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> 29 15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> 30 19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> 31 15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> 32 21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#> 
#> $at
#> $at$hp
#> [1] 146.6875
#> 
#> $at$wt
#> [1] 3.21725
#> 
#> 
#> $weights
#> [1] "proportional"

#> Coercing variables according to: $carb
#> [1] "numeric"
#> 

# update the confidence level used
# NOTE: if you're examining confidence intervals, it's much faster to udpate
# rather than re-run the model and means
sas.lsm <- update(sas.lsm, level = 0.95)

# using the means as a dataframe
as.data.frame(sas.lsm)

#>   carb carb.lsmeans.factor   lsmean        SE df lower.CL upper.CL .wgt.
#> 1    1                   1 20.63151 1.2682590 24 18.01395 23.24906     7
#> 2    2                   2 20.10695 0.9417013 24 18.16337 22.05053    10
#> 3    3                   3 19.74925 1.6608191 24 16.32148 23.17701     3
#> 4    4                   4 19.61794 1.0162966 24 17.52041 21.71548    10
#> 5    6                   6 19.11064 2.8433240 24 13.24231 24.97897     1
#> 6    8                   8 22.87216 3.7432900 24 15.14639 30.59793     1