Compute LSMEANS to mirror SAS functionality

Compute LSMEANS to mirror SAS functionality

sas.lsmeans.orig(model.obj, specs, data = NULL, mode = "kenward-roger",
  quietly = FALSE, verbose = FALSE, ...)

Arguments

model.obj
A model object on which to calculate lsmeans

specs
specs to be passed to lsmeans()

mode
mode to be passed to lsmeans()

quietly
whether to hide message informing of changes to default functionality

...
additional arguments to be passed to lsmeans

as.data.frame
whether to return a lsm.list object or a data.frame

Examples

library(dplyr) # for %>% operator

# create model on which to calculate lsmeans
gls(mpg ~ hp + carb + wt,
    data = mtcars %>%
             mutate(carb = as.factor(carb)), # make sure contrast is factor
    na.action = na.exclude) %>%

# call lsmeans, specifying factors over which means should be calculated
sas.lsmeans(~ carb)

#> Error in eval(expr, envir, enclos): could not find function "gls"